Improve handling of channel numbers in UI to prevent spurious errors
Change the GUI look a bit, making the "extra" options center-aligned with some padding.

Hopefully resolve the weirdness when the 3 color deconvolved channels would
count as selected, but not show up visually as selected, if switching from multiplex
to brightfield.

Change the handling of channels to read the channel information from the RDF spec
and enable/disable running/error messages appropriately.

Running models is disabled and a warning shown if:

- the currently selected model doesn't support arbitrary channel numbers
- the number of currently selected channels doesn't match the number the model wants

This means you can do funny stuff like:

- select 3 channels in a multiplexed image and run a brightfield model on it
- Run a brightfield model on the color deconvolved channels of an RGB image
- Run a fluorescence model on brightfield (color deconvolved or RGB)
- Support weird models like the 2 channel tissuenet one