Avoid storing the ImageData in InstanSeg
The InstanSeg object really encapsulates the stuff needed to run detection for an image - it does not need to contain the `ImageData` object itself.

This PR removes it so that the `ImageData` is passed at the time when object detection is called.

This should also fix a few bugs along the way, related to channels be specified in an unexpected way (e.g. `allChannels()` called before the `ImageData` had been set).

It also simplifies things slightly by passing the builder to the `InstanSeg` constructor, so that it can extract the fields that it wants - rather than needing to pass them as parameters.

Lastly, there doesn't really seem to be a need to treat `makeMeasurements` differently from other parameters passed to the builder, so now there is just one (overloaded) method to `detectObjects` that may or may not make measurements, depending upon whether it was requested in the builder.