Implement model versioning (#113)
Add a "local" model directory for local models.
Replace GitHub parsing with an index hosted within the extension
Add a version field for the InstanSegModel class.
  For remote models, this is set to be the release name
  For downloaded and local models, this is set to be the version in the rdf yaml (nullable).
Download models to model_name (from the rdf yaml), and versions in a subdirectory (eg 0.0.1)
Change the model "toString" so that it displays the model parent directory and name, followed by the version (if present)
  For local models, this is local/my_model-0.1.0. Can mismatch for local models but that's fine.
  For remote models, this will be eg my_model-0.0.1.

Downloading works as before, where we download zips on request and hopefully don't do anything twice, but still check for validity etc. This would work without needing to remove previous model versions, although you would get a weird structure where brightfield_nuclei/ contains 0.0.1 and 0.0.2 model directories, but also the files from the original 0.0.1 download.

Currently, this is implemented with the releases on my fork. There is a minor mismatch in that the rdf version is "0.0.1" while the release name is "v0.0.1", but this is easily fixable.

Also, log some more errors.

Resolves #67.