Support multi-output models with bioimageio spec 0.5 (#117)
Add support for multi-output models using the bioimageio 0.5.x model spec.

We currently handle logits and embeddings by adding them to the measurements of output objects. We don't have a good way to name measurements yet; hopefully we can get axis names at some point in the future of the bioimageio spec. Detection classes are set as PathClass on the output objects.

Detection classes are supported with a NominalOrOrdinalDataDescr. If logits are present too, we try to match the names of outputs and logits. Otherwise, we default to "Class 1", "Logit 1" etc.

Semantic segmentation is recognised as an output type, but support is not currently implemented (suggest this could be a future PR for now?).

Superset of #100

---------

Co-authored-by: Pete <pete.bankhead@gmail.com>