Improve channel selection on model download
Attempt to make a better selection of input and output channels when a model is first downloaded.

Before downloading, it's not known exactly what channels are available or required - and currently the channels may be valid when the user clicks 'Run', but then become reset to something invalid once the model arrives.

I don't think this convoluted code handles things perfectly, but I think it is an improvement... at least from my explorations, after repeatedly deleting models and redownloading across a few images.