Remove 'helpful' model selection logic and memory failure suggestion
As failures are often not related to memory, suggesting as
much may be harmful rather than helpful. Remove the suggestion
from the error message, with a troubleshooting page to be
considered for a later date (address #175).

Remove logic for automatically selecting models, as this
may lead to a brightfield model being selected for a fluorescence
image, or vice versa. Consider setting default model choices
for fluoro/brightfield imagedata at a later date (resolve #178).